+++ formula-android/src/main/java/com/instacart/formula/android/FragmentStore.kt
@@ -13,47 +13,76 @@ import io.reactivex.rxjava3.core.Observable
  * A FragmentStore is responsible for managing the state of multiple [FragmentKey] instances.
  */
 class FragmentStore @PublishedApi internal constructor(
+    val environment: FragmentEnvironment,
     private val formula: FragmentStoreFormula,
+    internal val onPreRenderFragmentState: ((FragmentState) -> Unit)? = null,
+    private val onFragmentLifecycleEvent: ((FragmentLifecycleEvent) -> Unit)? = null,
 ) {
-    companion object {
-        val EMPTY = init {  }
 
-        fun init(
-            init: FeaturesBuilder<Unit>.() -> Unit
-        ): FragmentStore {
-            return init(Unit, init)
+    class Builder {
+        private var environment: FragmentEnvironment? = null
+        private var onPreRenderFragmentState: ((FragmentState) -> Unit)? = null
+        private var onFragmentLifecycleEvent: ((FragmentLifecycleEvent) -> Unit)? = null
+
+        fun setFragmentEnvironment(environment: FragmentEnvironment) = apply {
+            this.environment = environment
+        }
+        
+        fun setOnPreRenderFragmentState(callback: ((FragmentState) -> Unit)?): Builder = apply {
+            this.onPreRenderFragmentState = callback
         }
 
-        fun <Component> init(
+        fun setOnFragmentLifecycleEvent(callback: ((FragmentLifecycleEvent) -> Unit)?): Builder = apply {
+            this.onFragmentLifecycleEvent = callback
+        }
+        
+        fun build(
+            init: FeaturesBuilder<Unit>.() -> Unit
+        ): FragmentStore {
+            return build(Unit, init)
+        }
+
+        fun <Component> build(
             rootComponent: Component,
             init: FeaturesBuilder<Component>.() -> Unit
         ): FragmentStore {
             val features = FeaturesBuilder.build(init)
-            return init(rootComponent, features)
+            return build(rootComponent, features)
         }
 
-        fun <Component> init(
+        fun <Component> build(
             component: Component,
             features: Features<Component>
         ): FragmentStore {
             val featureComponent = FeatureComponent(component, features.bindings)
-            val formula = FragmentStoreFormula(featureComponent)
-            return FragmentStore(formula)
+            val fragmentEnvironment = environment ?: FragmentEnvironment()
+            val formula = FragmentStoreFormula(fragmentEnvironment, featureComponent)
+            return FragmentStore(
+                environment = fragmentEnvironment,
+                formula = formula,
+                onPreRenderFragmentState = onPreRenderFragmentState,
+                onFragmentLifecycleEvent = onFragmentLifecycleEvent
+            )
         }
     }
 
-    internal fun onLifecycleEffect(event: FragmentLifecycleEvent) {
+    companion object {
+        val EMPTY = Builder().build {  }
+    }
+
+    internal fun onLifecycleEvent(event: FragmentLifecycleEvent) {
         formula.onLifecycleEffect(event)
+        onFragmentLifecycleEvent?.invoke(event)
     }
 
     internal fun onVisibilityChanged(fragmentId: FragmentId, visible: Boolean) {
         formula.onVisibilityChanged(fragmentId, visible)
     }
 
-    internal fun state(environment: FragmentEnvironment): Observable<FragmentState> {
+    internal fun state(): Observable<FragmentState> {
         val config = RuntimeConfig(
             defaultDispatcher = MainThreadDispatcher(),
         )
-        return formula.toObservable(environment, config)
+        return formula.toObservable(config)
     }
 }