+++ formula-android/src/main/java/com/instacart/formula/android/internal/FragmentFlowRenderView.kt
@@ -15,6 +15,7 @@ import com.instacart.formula.android.events.FragmentLifecycleEvent
 import com.instacart.formula.android.BackCallback
 import com.instacart.formula.android.FeatureEvent
 import com.instacart.formula.android.FragmentId
+import com.instacart.formula.android.FragmentStore
 import com.instacart.formula.android.ViewFactory
 import com.instacart.formula.android.getFormulaFragmentId
 import java.util.LinkedList
@@ -25,19 +26,19 @@ import java.util.LinkedList
  * NOTE: Initialize this class before calling [FragmentActivity.super.onCreate]
  *
  * [activity] activity within which the [FragmentFlowRenderView] lives.
- * [onLifecycleEvent] fragment lifecycle events that should be passed to the [com.instacart.formula.android.FragmentStore]
  */
 internal class FragmentFlowRenderView(
     private val activity: FragmentActivity,
-    private val fragmentEnvironment: FragmentEnvironment,
-    private val onLifecycleEvent: (FragmentLifecycleEvent) -> Unit,
+    private val store: FragmentStore,
     private val onLifecycleState: (FragmentId, Lifecycle.State) -> Unit,
     private val onFragmentViewStateChanged: (FragmentId, isVisible: Boolean) -> Unit
 ) {
-
     private var fragmentState: FragmentState? = null
     private var features: Map<FragmentId, FeatureEvent> = emptyMap()
     private val visibleFragments: LinkedList<Fragment> = LinkedList()
+
+    private val environment: FragmentEnvironment
+        get() = store.environment
 
     private val featureProvider = object : FeatureProvider {
         override fun getFeature(id: FragmentId): FeatureEvent? {
@@ -91,13 +92,18 @@ internal class FragmentFlowRenderView(
 
         override fun onFragmentAttached(fm: FragmentManager, f: Fragment, context: Context) {
             super.onFragmentAttached(fm, f, context)
+            if (f is FormulaFragment) {
+                f.environment = environment
+            }
+
             if (FragmentLifecycle.shouldTrack(f)) {
                 val event = FragmentLifecycleEvent.Added(
                     fragmentId = f.getFormulaFragmentId(),
                 )
-                onLifecycleEvent(event)
+
+                store.onLifecycleEvent(event)
             } else {
-                fragmentEnvironment.logger("Ignoring attach event for fragment: $f")
+                environment.logger("Ignoring attach event for fragment: $f")
             }
         }
 
@@ -111,7 +117,7 @@ internal class FragmentFlowRenderView(
                     fragmentId = f.getFormulaFragmentId(),
                     lastState = formulaFragment?.currentState(),
                 )
-                onLifecycleEvent(event)
+                store.onLifecycleEvent(event)
             }
         }
     }
@@ -122,6 +128,8 @@ internal class FragmentFlowRenderView(
 
     fun render(state: FragmentState) {
         Utils.assertMainThread()
+
+        store.onPreRenderFragmentState?.invoke(state)
 
         fragmentState = state
         features = state.features
@@ -139,7 +147,7 @@ internal class FragmentFlowRenderView(
 
     fun viewFactory(fragment: FormulaFragment): ViewFactory<Any> {
         return FormulaFragmentViewFactory(
-            environment = fragmentEnvironment,
+            environment = environment,
             fragmentId = fragment.getFormulaFragmentId(),
             featureProvider = featureProvider,
         )