+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - **Breaking**: Move factory functions from `FlowAction` to `Action`
 - **Breaking**: Move coroutines logic into `formula` and removed `formula-coroutines` module
 - **Breaking**: Remove `ActivityStore.streams`
+- **Breaking**: Moving `FragmentEnvironment` to `FragmentStore`
 
 ## [0.7.1] - June 28, 2022
 - **Breaking**: Rename `FragmentBindingBuilder` to `FragmentStoreBuilder`